requirements.txt
sentence-transformers
faiss-cpu
PyMuPDF
tqdm
